move inputs to shared/ui folder
