+++ packages/hash/frontend/src/shared/ui.ts
@@ -5,6 +5,7 @@ export * from "./ui/link";
 export * from "./ui/loading-spinner";
 export * from "./ui/popover";
 export * from "./ui/submit-button-wrapper";
+export * from "./ui/text-field";
 
 export * from "./ui/theme";
 export * from "./ui/utils";
