+++ packages/hash/frontend/src/shared/ui/theme/components/inputs.ts
@@ -1,2 +1,6 @@
 export * from "./inputs/mui-button-theme-options";
 export * from "./inputs/mui-outlined-input-theme-options";
+export * from "./inputs/mui-checkbox-theme-options";
+export * from "./inputs/mui-input-label-theme-options";
+export * from "./inputs/mui-radio-theme-options";
+export * from "./inputs/mui-switch-theme-options";
