+++ packages/hash/frontend/src/shared/ui/theme/components/inputs/mui-checkbox-theme-options.tsx
@@ -1,22 +1,26 @@
-import { Components, Theme } from "@mui/material";
-import { RadioCheckedIcon } from "../icons/RadioCheckedIcon";
-import { RadioUncheckedIcon } from "../icons/RadioUncheckedIcon";
+import { Components, Theme, svgIconClasses } from "@mui/material";
+import { CheckboxBlankIcon } from "./mui-checkbox-theme-options/checkbox-blank-icon";
+import { CheckboxIcon } from "./mui-checkbox-theme-options/checkbox-icon";
 
 const focusBorderOffset = 4;
 const focusBorderWidth = 2;
+const checkboxBorderRadius = 3;
 
-export const MuiRadioThemeOptions: Components<Theme>["MuiRadio"] = {
+export const MuiCheckboxThemeOptions: Components<Theme>["MuiCheckbox"] = {
   defaultProps: {
-    disableRipple: true,
     disableFocusRipple: true,
+    disableRipple: true,
     disableTouchRipple: true,
-    icon: <RadioUncheckedIcon />,
-    checkedIcon: <RadioCheckedIcon />,
+    icon: <CheckboxBlankIcon />,
+    checkedIcon: <CheckboxIcon />,
   },
   styleOverrides: {
     root: ({ theme }) => ({
-      color: theme.palette.gray[40],
       padding: 0,
+
+      [`.${svgIconClasses.root}`]: {
+        position: "relative",
+      },
 
       [`&.Mui-focusVisible:after, &:focus-visible:after`]: {
         content: `""`,
@@ -28,7 +32,7 @@ export const MuiRadioThemeOptions: Components<Theme>["MuiRadio"] = {
         bottom: -focusBorderOffset,
         borderWidth: focusBorderWidth,
         borderColor: theme.palette.blue[70],
-        borderRadius: "50%",
+        borderRadius: checkboxBorderRadius + focusBorderOffset,
       },
     }),
   },
