+++ packages/hash/frontend/src/shared/ui/theme/components/inputs/mui-checkbox-theme-options/checkbox-blank-icon.tsx
