+++ packages/hash/frontend/src/shared/ui/theme/components/inputs/mui-radio-theme-options/radio-unchecked-icon.tsx
