+++ packages/hash/frontend/src/shared/ui/theme/components/inputs/mui-switch-theme-options.ts
